+++ modules/target.lua
@@ -500,6 +500,10 @@ PlayerTargetChanged:SetScript("OnEvent", function(self, event)
             local info = frame.BetterBlizzPlates.unitInfo
             frame.BetterBlizzPlates.unitInfo = BBP.GetNameplateUnitInfo(frame)
 
+            if config.healthNumbersTargetOnly then
+                BBP.HealthNumbers(frame)
+            end
+
             -- info.isTarget = false
             -- info.wasTarget = true
 
@@ -567,6 +571,10 @@ PlayerTargetChanged:SetScript("OnEvent", function(self, event)
                 if config.fadeAllButTarget then
                     BBP.FadeAllButTargetNameplates()
                 end
+            end
+
+            if config.healthNumbersTargetOnly then
+                BBP.HealthNumbers(frame)
             end
 
             if config.hideNPC then BBP.HideNPCs(frame, targetNameplate) end